5.0-rag-documentation-helper/pyproject.toml
[project]
name = "5-0-rag-documentation-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-openai>=1.2.1",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.18",
    "langchain-text-splitters>=1.1.2",
    "pinecone[grpc]>=7.3.0",
    "python-dotenv>=1.2.2",
]
